fix bug with large ASN4 (value) in as-path

+++ lib/exabgp/bgp/message/open/asn.py
@@ -11,7 +11,7 @@ from struct import pack
 
 # =================================================================== ASN
 
-class ASN (int):
+class ASN (long):
 	def asn4 (self):
 		return self > pow(2,16)
